+++ app/grails-app/controllers/com/k_int/kbplus/SubscriptionController.groovy
@@ -922,7 +922,7 @@ class SubscriptionController extends AbstractDebugController {
         def message = escapeService.escapeString(result.subscription.name) + "_" + g.message(code: 'subscriptionDetails.members.members') + "_" + datetoday
         def orgs = []
         if (params.exportXLS || params.format) {
-            Map allContacts = Person.getPublicAndPrivateEmailByFunc('General contact person')
+            Map allContacts = Person.getPublicAndPrivateEmailByFunc('General contact person',result.institution)
             Map publicContacts = allContacts.publicContacts
             Map privateContacts = allContacts.privateContacts
             result.filteredSubChilds.each { row ->
@@ -942,12 +942,12 @@ class SubscriptionController extends AbstractDebugController {
                     org.status = subChild.status
                     org.customProperties = subscr.customProperties
                     org.privateProperties = subscr.privateProperties
-                    String generalContacts = ""
+                    Set generalContacts = []
                     if (publicContacts.get(subscr))
-                        generalContacts += publicContacts.get(subscr).join("; ") + "; "
+                        generalContacts.addAll(publicContacts.get(subscr))
                     if (privateContacts.get(subscr))
-                        generalContacts += privateContacts.get(subscr).join("; ")
-                    org.generalContacts = generalContacts
+                        generalContacts.addAll(privateContacts.get(subscr))
+                    org.generalContacts = generalContacts.join("; ")
                     orgs << org
                 }
             }
@@ -1869,7 +1869,7 @@ class SubscriptionController extends AbstractDebugController {
                                 //name: result.subscriptionInstance.name + " (" + (cm.get(0).shortname ?: cm.get(0).name) + ")",
                                 startDate: startDate,
                                 endDate: endDate,
-                                administrative: result.subscriptionInstance.administrative,
+                                administrative: result.subscriptionInstance.getCalculatedType() == TemplateSupport.CALCULATED_TYPE_ADMINISTRATIVE,
                                 manualRenewalDate: result.subscriptionInstance.manualRenewalDate,
                                 /* manualCancellationDate: result.subscriptionInstance.manualCancellationDate, */
                                 identifier: java.util.UUID.randomUUID().toString(),
@@ -1892,7 +1892,7 @@ class SubscriptionController extends AbstractDebugController {
 
                         if (cons_sub) {
 
-                            if(cons_sub.administrative)
+                            if(cons_sub.getCalculatedType() == TemplateSupport.CALCULATED_TYPE_ADMINISTRATIVE)
                                 new OrgRole(org: cm, sub: cons_sub, roleType: role_sub_hidden).save()
                             else
                                 new OrgRole(org: cm, sub: cons_sub, roleType: role_sub).save()
@@ -3664,6 +3664,21 @@ AND l.status.value != 'Deleted' AND (l.instanceOf is null) order by LOWER(l.refe
         switch (params.workFlowPart) {
             case WORKFLOW_DATES_OWNER_RELATIONS:
                 result << copySubElements_DatesOwnerRelations();
+                if (params.isRenewSub){
+                    result << loadDataFor_PackagesEntitlements()
+                }
+                break;
+            case WORKFLOW_SUBSCRIBER:
+                result << copySubElements_Subscriber();
+//                if (params.isRenewSub) {
+//                    result << loadDataFor_Subscriber()
+//                }
+                break;
+            case WORKFLOW_PACKAGES_ENTITLEMENTS:
+                result << copySubElements_PackagesEntitlements();
+                if (params.isRenewSub){
+                    result << loadDataFor_DocsAnnouncementsTasks()
+                }
                 break;
             case WORKFLOW_DOCS_ANNOUNCEMENT_TASKS:
                 result << copySubElements_DocsAnnouncementsTasks();
@@ -3671,17 +3686,11 @@ AND l.status.value != 'Deleted' AND (l.instanceOf is null) order by LOWER(l.refe
                     result << loadDataFor_Properties()
                 }
                 break;
-            case WORKFLOW_SUBSCRIBER:
-                result << copySubElements_Subscriber();
-                break;
             case WORKFLOW_PROPERTIES:
                 result << copySubElements_Properties();
                 if (params?.targetSubscriptionId && params.isRenewSub){
                     redirect controller: 'subscription', action: 'show', params: [id: params?.targetSubscriptionId]
                 }
-                break;
-            case WORKFLOW_PACKAGES_ENTITLEMENTS:
-                result << copySubElements_PackagesEntitlements();
                 break;
             default:
                 result << loadDataFor_DatesOwnerRelations()
@@ -3845,6 +3854,31 @@ AND l.status.value != 'Deleted' AND (l.instanceOf is null) order by LOWER(l.refe
             params.workFlowPart = WORKFLOW_DOCS_ANNOUNCEMENT_TASKS
             params.workFlowPartNext = WORKFLOW_PROPERTIES
         }
+
+        result.sourceSubscription = baseSub
+        result.targetSubscription = newSub
+        result.sourceTasks = taskService.getTasksByResponsiblesAndObject(result.user, contextService.org, result.sourceSubscription)
+        result.targetTasks = taskService.getTasksByResponsiblesAndObject(result.user, contextService.org, result.targetSubscription)
+        result
+    }
+    private loadDataFor_DocsAnnouncementsTasks() {
+        def result = setResultGenericsAndCheckAccess(AccessService.CHECK_VIEW)
+        Subscription baseSub = Subscription.get(params.sourceSubscriptionId ? Long.parseLong(params.sourceSubscriptionId): Long.parseLong(params.id))
+        Subscription newSub = null
+        if (params.targetSubscriptionId) {
+            newSub = Subscription.get(Long.parseLong(params.targetSubscriptionId))
+        }
+//
+//        if (isTargetSubChanged) {
+//            newSub = newSub.refresh()
+//        }
+//        if (params.isRenewSub) {
+//            params.workFlowPart = WORKFLOW_PROPERTIES
+//            params.workFlowPartNext = WORKFLOW_END
+//        }else {
+//            params.workFlowPart = WORKFLOW_DOCS_ANNOUNCEMENT_TASKS
+//            params.workFlowPartNext = WORKFLOW_PROPERTIES
+//        }
 
         result.sourceSubscription = baseSub
         result.targetSubscription = newSub
@@ -4072,6 +4106,28 @@ AND l.status.value != 'Deleted' AND (l.instanceOf is null) order by LOWER(l.refe
             params.workFlowPart = WORKFLOW_PACKAGES_ENTITLEMENTS
             params.workFlowPartNext = WORKFLOW_DOCS_ANNOUNCEMENT_TASKS
         }
+
+        result.sourceIEs = subscriptionService.getIssueEntitlements(baseSub)
+        result.targetIEs = subscriptionService.getIssueEntitlements(newSub)
+        result.newSub = newSub
+        result.subscription = baseSub
+        result
+    }
+    private loadDataFor_PackagesEntitlements() {
+        def result = setResultGenericsAndCheckAccess(AccessService.CHECK_VIEW)
+        Subscription baseSub = Subscription.get(params.sourceSubscriptionId ?: params.id)
+        Subscription newSub = params.targetSubscriptionId ? Subscription.get(params.targetSubscriptionId) : null
+
+//        if (isTargetSubChanged) {
+//            newSub = newSub.refresh()
+//        }
+//        if (newSub && params.isRenewSub) {
+//            params.workFlowPart = WORKFLOW_DOCS_ANNOUNCEMENT_TASKS
+//            params.workFlowPartNext = WORKFLOW_PROPERTIES
+//        }else {
+//            params.workFlowPart = WORKFLOW_PACKAGES_ENTITLEMENTS
+//            params.workFlowPartNext = WORKFLOW_DOCS_ANNOUNCEMENT_TASKS
+//        }
 
         result.sourceIEs = subscriptionService.getIssueEntitlements(baseSub)
         result.targetIEs = subscriptionService.getIssueEntitlements(newSub)