+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.17.5
+
+- fixed: tenant check for private contacts
+- fixed: default dashboard and dashboard tab for new users
+- added information for ip-ranges
+
 0.17.4
 
 - improvement org profil